xerrors: redirect to go1.13 primitives for 1.13
For interoperability it is especially important that the
Printer type is redirected. Otherwise the FormatError method
would not be of the same signature and there would be
no interoperability.

Some of the more bulky code is now also redirected to use
1.13.

Tests are verified to pass for both go1.12 and go1.13.

Change-Id: Ia5776252aa4774d900c53661659d9a8b61793015
Reviewed-on: https://go-review.googlesource.com/c/xerrors/+/167577
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Damien Neil <dneil@google.com>